Add "StackPieceTracers"
This commit overhauls how the Parser callbacks work in the Interpreter.

The main problem is this: How do we keep track of things like undefined
identifiers? The tokens have the data necessary to generate
TracebackErrors. However, we need a way to keep track of the tokens that
match their corresponding engine objects.

This is exactly why the StackPieceTracer class was created.
It allows (and enforces) engine objects being paired with a list of
tokens. On creation, tokens are required to pair to an object, and a
trace() function is given to keep track of tokens we need to check later

As a minor optimization, StackPieceTracers can "append" new data.
Instead of creating a new stack piece every time objects are combined
(which also loses track of previously made traces), stack pieces utilize
update(), a function that swaps the old object with a new object, as
well as swapping its paired tokens. On top of that, an extra list can
(er, must) be given in the cases where stack piece objects combine into
one engine object.

Now we can track traces of tokens through the parsing stack!