Implement "operation lists" to fix Expression eval
When given the expression 3 - 4 + 2, the interpreter evaluates
right-to-left (because the parser does). This is not how the expression
should be evaluated.

To compensate, stacks were added at every level of
precedence for operation productions, which generate a list of all
expressions of the same precedence.
In the next production up (the one with the next lowest precedence),
this list can be popped and evaluated from left-to-right as intended,
and the resulting expression used in another operation list.

With a stroke of luck, an "unneeded" expression production in the parser
allows us to hook into the operation list that onOperationLow() pushes
to its stack. By using the onExpression() callback, we are able to
evaluate that last operation list into a single expression to push to
the "expressions" stack for the calling production. Turns out that extra
single-branch "expression" production was handy after all!

(This commit also includes a minor change to the triple-negative
interpreter test. This was done to test the recently implemented
Expression.__str__() method)