Improve expressables and engine key-testing
TemplateCalls need to be comparable if they are going to be traced by
the interpreter. For this reason, all Expressables are now comparable
and hashable.

With this change, the Engine class also optimizes checking for
the required functions (for classes that can be used as dict keys) by
checking in __init__ instead of every setAlias() call