Implement Containable mixin (and modify heirarchy)
This commit changes the mixins and most data structues by adding the
"Containable" type. This type guarantees a function that asks if an
identifier is "contained" in the data structure.

As an added bonus, wrapper functions were defined to check types of
returned values in abstract methods. This, of course, meant changing the
function names in all the Engine classes, but now it should be easier to
find errors when they arise in implementing relations