Add relation creation handling to SolverPro
This commit is another big one. First, the Engine module needed to
complete the Solutions() class, and eventually a SolutionSet was created
to help manage the results of sympy.solveset().

The Engine itself was modified as well; Engine._identifiers is now a
list(), and _aliases (dict) and _solutions (list) were added as well.
(This will help when commands are implemented and the order of
identifier definitions will be important.)

The Interpreter added functionality to parse relations and add them to
the engine. On top of that, a new error type was introduced to handle
cases where a template identifier was evaluated in some part of an
expression. This bloated the expression branch of executeLine(), but
under the current formatting of Interpreter, it is as concise as
possible.

A new test case was also added to ensure the Interpreter prints an error
when a template identifier is treated as a numeric one.