Add separate template substitution
This commit adds the ability to substitute template calls only. This
allows the interpreter to evaluate template aliases when the relations
are created.

A test was created to ensure this behavior. It creates a template alias,
evaluates a new relation, changes the template alias, and then evaluates
the relation. The expected output is the return value of the old
template alias. This ensures that relations do not evaluate template
aliases "on the fly", and instead are created with expressions that have
their template aliases already evaluated.