Improve type tree for interpreter errors
This commit changes the InterpreterError class, and makes it a new type
otherwise identical to Exceptions. This allowed the
InterpreterNotImplementedError type to be declared (it isn't a traceback
error.)

The above was done to allow the Interpreter to distinguish between
"interpreter not implemented errors" and "regular not implemented
errors".

To reintroduce the TracebackError functionality,
InterpreterTracebackError and InterpreterTracebackWarning were added,
with their child classes inheriting from these now.

Now, whenever any InterpreterError is thrown, it is considered "handled"
by the interpreter. Any other Exception() thrown will be treated as an
"internal error".