Implement (correct) structures in Interpreter
This commit is HUGE... I'll try to break it down

First off, Structures have now been added. They are (hopefully) much
more organized and useful compared to the other versions in engine.py.
The main differences to note are:
- Expression() signature has changed
    (oper first, then variable number of args)
- Mixins have more well-defined guarantees
- Sets have been removed (for now)

The interpreter now has two stages for ensuring valididity. The first is
the "REPRESENT" stage (denoted by _REP on the tail end of ensure-fns).
The second is the "TEMPLATE" stage (denoted by _TEMP on the tail end of
ensure-fns). Separation of these stages allows representations to be
evaluated before proceeding to evaluating template calls, allowing the
actual TemplateCall() to be used with its builtin functionality (instead
of re-writing all of it for Represent() objects).

It is worth noting that TemplateCall()s are necessary to have because
Template()s do not evaluate immediately, and they need a way to
dynamically represent an evaluation of a template alias.

Represent() objects now implement __repr__ (for debugging)

Another small change to note is the interpreter now ignores comments
(indeed, it would have errored this whole time if given some).

A new trace type was added to aid in checking for bad template calls.
These checks aren't finished, however.

The database had a few bug fixes, as well as more efficient code dealing
with generators.

All in all, Interpreter should be working at this point, although it
does not yet have the ability to process relations (and onValue()
throws for template calls, due to them not being completely finished).