Optimize recursiveSolve()
This is a sad commit... recursiveSolve() had a bug where variables were
not getting substituted. Because of that bug, database.substitute()
has a bunch of extra "solving code". Thinking about it more, though, it
makes sense that substitute() works the way it does, however
recursiveSolve() never really did what it was intended to (especially
now that "the intentions" have changed).

At this point, recursiveSolve() effectively only collapsed all solutions
into a dictionary of variables to solution lists. This commit optimizes
(and renames) recursiveSolve() to do what it actually does.