Overhaul (and complete) database.substitute()
(A few TODOs were added as well...)

This commit finishes the substitute() function... after overhauling it
AGAIN. But hey! It should be faster now that it is broken up into a
couple helper iterators, and it utilizes sympy's (probably)
more efficient functionality. In fact... I'm not sure there is any use
for the structures.py file. That should be removed at some point...