Implement Relations (#1)
* Move immutable decorator to standalone file

* Add InterpreterDatabase and FilteredDict

* Add a wrapper to parser using InterpreterParser

* Modify/finish InterpreterDatabase and FilteredDict

(FilteredDict now has updatable filters)

* Move interpreter errors to errors.py

* Make InterpreterParser use Represent()

This commit changes the InterpreterParser to utilize a cleaner
communication method with the Interpreter. Ultimately, parsing should be
just that: parsing. No data structures (like Identifier or Expression)
should be created. Representing them in an easy-to-evaluate way allows
the interpreter to check and print errors before making the objects

* Remove template filter from database

With the new Represent() change, this allows the Interpreter to
substitute template calls before making any other evaluations (making
these functions/features unneeded)

* Add/improve tests

* Finish refactoring Interpreter (for now...)

Please note this commit as well as all before this are not properly
functioning code changes. A future commit will provide a working version

* Make map() calls more concise in some cases

* Fix tests to allow abstract error types to work

* Remove main runline (no more evaluateLine!)

* Add a few more tests

Added tests for comments
Added more tests for detecting errors in template calls

* Imrpove errors

This commit does quite a few things:
- Makes some comments more helpful
- TracebackError is now an abstract class
- InterpreterNotImplementedError inherits NotImplementedError
- TracebackErrors indent multi-line messages
- map() is no longer used
- Errors are in-order according to when their corresponding checks run

* Fixed a small issue with lexer

* Implement (correct) structures in Interpreter

This commit is HUGE... I'll try to break it down

First off, Structures have now been added. They are (hopefully) much
more organized and useful compared to the other versions in engine.py.
The main differences to note are:
- Expression() signature has changed
    (oper first, then variable number of args)
- Mixins have more well-defined guarantees
- Sets have been removed (for now)

The interpreter now has two stages for ensuring valididity. The first is
the "REPRESENT" stage (denoted by _REP on the tail end of ensure-fns).
The second is the "TEMPLATE" stage (denoted by _TEMP on the tail end of
ensure-fns). Separation of these stages allows representations to be
evaluated before proceeding to evaluating template calls, allowing the
actual TemplateCall() to be used with its builtin functionality (instead
of re-writing all of it for Represent() objects).

It is worth noting that TemplateCall()s are necessary to have because
Template()s do not evaluate immediately, and they need a way to
dynamically represent an evaluation of a template alias.

Represent() objects now implement __repr__ (for debugging)

Another small change to note is the interpreter now ignores comments
(indeed, it would have errored this whole time if given some).

A new trace type was added to aid in checking for bad template calls.
These checks aren't finished, however.

The database had a few bug fixes, as well as more efficient code dealing
with generators.

All in all, Interpreter should be working at this point, although it
does not yet have the ability to process relations (and onValue()
throws for template calls, due to them not being completely finished).

* Favor generator expressions over map()

* Add test for future feature

* Remove engine.py (there's a copy in git stash)

* Fix a lexer bug with numbers (WHAT?!)

Yep! I found a bug in my LEXER! It wouldn't register "5." as a number,
because of a "required digit after the period"

* Add ability to store relations

This commit adds the ability for the Interpreter to add/process
relations. The Interpreter can now correctly solve relations and add
the solutions to the database.

(This commit also contains future changes to Represent() regarding
changes to structure classes)

* Make several changes to main structures

1. FactoryRecipe, TemplateCall, and Expression no longer take
    variable-number arguments
2. Add solve() and solveAll() to Relation
3. Remove all notions of "StructureSet"s
4. Fix bug is-comparing type(class) to type (with ABC, the type is
    ABC.meta)

* Fully implement Interpreter ability to substitute

Man, this one was a PAIN! InterpreterDatabase now has substitute() and
_substituteUntilNumeric(). Several designs for each were considered,
and eventually the recursive _substituteUntilNumeric() settled on
"I return None for cyclical definitions, otherwise I return the set of
substitutions with as many numerics as possible (unless no numerics were
found, then I return the original expr)

There are several comments in the function explaining why these choices
were made, but the result passed the tests!

* Fix minor bug in Expressable algebra

(Created an invalid Expression object)

* Add a TODO

* Fix small FactoryRecipe bug

* Fix another TemplateCall bug

* Add ability to process sqrt() expressions

* Add test for simplification logic

* Fix precedence checking in Expression()s

* Optimize recursiveSolve()

This is a sad commit... recursiveSolve() had a bug where variables were
not getting substituted. Because of that bug, database.substitute()
has a bunch of extra "solving code". Thinking about it more, though, it
makes sense that substitute() works the way it does, however
recursiveSolve() never really did what it was intended to (especially
now that "the intentions" have changed).

At this point, recursiveSolve() effectively only collapsed all solutions
into a dictionary of variables to solution lists. This commit optimizes
(and renames) recursiveSolve() to do what it actually does.

* Implement better printing

* Overhaul (and complete) database.substitute()

(A few TODOs were added as well...)

This commit finishes the substitute() function... after overhauling it
AGAIN. But hey! It should be faster now that it is broken up into a
couple helper iterators, and it utilizes sympy's (probably)
more efficient functionality. In fact... I'm not sure there is any use
for the structures.py file. That should be removed at some point...

* Modify print fn (again) and help Sympy init

* Un-implement template calls (for version release!)

* Remove useless TODO

* Fix an incorrect error message

* New docs structure and work on some

This commit makes/modifies/renames some files to make the docs folder
clearer. It also contains the finished (for now) verions of the examples
and features file, with some progress on the errors file.

* Add some template alias tests

* Finish documenting errors (and rename one)

Removed "Call" from "TemplateCallMismatchError"

* (Finish renaming for last commit)

* Restrict template definitions instead of calls