Improve ExpressionRepresentation
This commit specifically adds static methods to the class for wrapping
values in ExpressionRepresentations, whether they be Representation
objects or other pre-processed objects.

In order to do this, a DummyRepresentation type was added, which
essentially always "constructs" to the exact value it was created with
originally.

This commit also includes some improvements on assertions.