Fix exponent evaluation direction
This was done by converting the notion of an "opList" into its own OpList type.
OpLists allow pusing values alternating between ExprReps and OperReps, then
evaluating its "stack" just as was done previously. The added
functionality for exponents lies within the reverseEvalDirection(),
which will take the "rightmost" 3 Representations and evaluate instead
of the default "leftmost" Representations. This allows exponents to be
evaluated right-to-left, contrary to most cases.