Implement solving some exponent cases
This commit adds a whole lot of stuff that was needed to solve relations
with exponent expression terms. Here's a brief rundown:
- toNumber() was added to show intent to just convert something to a
number (either int or float)
- class NumericSymbol() provides a type to use in ConditionSets (since
they require symbols yet we want to solve for numerics...) We will have
to be careful to convert it back in future use cases.
- _solveExponent() is the bulk of the change, manually splitting an
expression into base parts, then using the dirtier sympy.solve() to
generate a solution. Of course, sympy.solve() returns a list of
solutions, which brings us to the implementation of...
- Branched-solutions; they represent a joined set of Solutions, which
should all be parsed and SubSet.join()-ed by whatever extractor is
currently using them/
- Solutions can now reverse a substitution that might have been made
previously. This is primarily needed for numerics to be substituted to
NumericSymbols to make them compatible with ConditionSets
- (Other less significant changes not noted for sake of brevity)